requirements.txt
fastapi
huggingface-hub
motor
numpy
pinecone-client
pydantic
pydantic_core
python-dotenv
sentence-transformers
torch
transformers
typing_extensions
uvicorn